Rename PrimitiveHasher => PrimitiveValueByteArrayConverter
